+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-05-05
+
 ### Added
 
 - Add Workbench as requirement for Clutter pieces missing a workstation requirement.
@@ -53,7 +55,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.3.0...HEAD
+[unreleased]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.1.1...0.1.2